fix: relax lrclib decoding for Apple Music imports

+++ ios/singword/Networking/LrclibLyricsDataSource.swift
@@ -5,7 +5,7 @@ private struct LrclibTrack: Decodable {
     let trackName: String?
     let artistName: String?
     let albumName: String?
-    let duration: Int?
+    let duration: Double?
     let plainLyrics: String?
     let syncedLyrics: String?
 }
